# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'laravel + dingoApi デモ'
  contact:
    name: hirose.morisaki
    url: /
    email: hirose.morisaki@gmail.com
  version: '1.0'
servers:
  -
    url: 'http://shopapi.test/api'
paths:
...
  '/admin/users/{id}':
    get:
      tags:
        - Admin
      summary: 会員詳細
      description: 会員詳細
      operationId: usersid
      parameters:
        -
          name: id
          in: path
          description: ユーザID
          required: true
          schema:
            type: number
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                oneOf:
                 - $ref: '#components/schemas/UserOfOne'
        # Definition of all error statuses
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        # 400:
        #   description: Bad request. User ID must be an integer and larger than 0.
        # 401:
        #   description: Authorization information is missing or invalid.
        # 404:
        #   description: A user with the specified ID was not found.
        # 5XX:
        #   description: Unexpected error.
      deprecated: false
      security:
        -
          bearer_token: []
  '/admin/users/{userid}/lock':
    patch:
      tags:
        - Admin
      summary: 会員ロック/非ロック
      description: 会員ロック/非ロック
      operationId: userlock
      parameters:
        -
          name: userid
          in: path
          description: ユーザID
          required: true
          schema:
            type: integer
      responses:
        204:
          description: noContent!
      deprecated: false
      security:
        -
          bearer_token: []
...
components:
  securitySchemes:
    Bearer:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header
      scheme: bearer_token
    passport:
      type: oauth2
      description: 'Laravel passport oauth2 security.'
      in: header
      scheme: https
      flows:
        password:
          authorizationUrl: 'http://localhost/oauth/authorize'
          tokenUrl: 'http://localhost/oauth/token'
          refreshUrl: 'http://localhost/token/refresh'
          scopes: []
    bearer_token:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header
  schemas:
...
    Error:
      type: object
      properties:
        status_code:
          type: integer
          example: 422
        message:
          type: string
          example: 'error message'
        errors:
          type: object
      required:
        - code
        - message
security:
  - []